reorganize router/indexer files
